+++ data/repository/src/main/java/com/freshdigitable/udonroad2/data/impl/UserRepository.kt
@@ -1,14 +1,14 @@
 package com.freshdigitable.udonroad2.data.impl
 
-import com.freshdigitable.udonroad2.data.UserRepository
+import com.freshdigitable.udonroad2.data.UserDataSource
 import com.freshdigitable.udonroad2.model.UserId
 import com.freshdigitable.udonroad2.model.user.UserEntity
 import kotlinx.coroutines.flow.Flow
 
-internal class UserRepositoryImpl(
-    private val localSource: UserRepository.LocalSource,
-    private val restClient: UserRepository.RemoteSource,
-) : UserRepository {
+internal class UserRepository(
+    private val localSource: UserDataSource.Local,
+    private val restClient: UserDataSource.Remote,
+) : UserDataSource {
     override fun getUserSource(id: UserId): Flow<UserEntity?> = localSource.getUserSource(id)
 
     override suspend fun findUser(id: UserId): UserEntity? = localSource.findUser(id)
